Icons updated, remove student enabled, ui improved.
*The done, and cancel icons have been modified.

*The option to remove a student is now working perfectly.

*The ui have been improved, specially the alert when a student have been edited.

*There's a bug when the user tries to remove all the signatures from the array of extra subjects.

+++ jfx-school-manager/src/model/StudentInclusion.java
@@ -4,7 +4,7 @@ import java.util.ArrayList;
 
 public class StudentInclusion extends Student{
 
-    public StudentInclusion(String pHasRelatives, long pCost, String pName, String pLastName, Courses pCourse, String pId, String pTerapy, ArrayList<ExtraAsignatures> pAsignatures) {
+    public StudentInclusion(String pHasRelatives, Float pCost, String pName, String pLastName, Courses pCourse, String pId, String pTerapy, ArrayList<ExtraAsignatures> pAsignatures) {
         super(pName, pLastName, pCourse, pHasRelatives, pCost, pTerapy, pId, pAsignatures);
     }
